add preferences component
