+++ mage_ai/api/presenters/ProjectPresenter.py
@@ -6,6 +6,7 @@ class ProjectPresenter(BasePresenter):
         'help_improve_mage',
         'latest_version',
         'name',
+        'openai_api_key',
         'project_uuid',
         'version',
     ]
