+++ mage_ai/frontend/interfaces/ProjectType.ts
@@ -2,6 +2,7 @@ export default interface ProjectType {
   help_improve_mage?: boolean;
   latest_version: string;
   name: string;
+  openai_api_key?: string;
   project_uuid?: string;
   version: string;
 }
